fix: resolve jest version conflict
